Separate the form part on EditEntity
