+++ package.json
@@ -22,6 +22,7 @@
   "homepage": "https://github.com/userlocalhost/airone#readme",
   "devDependencies": {
     "@babel/core": "^7.12.10",
+    "@babel/plugin-transform-runtime": "^7.14.5",
     "@babel/preset-env": "^7.12.11",
     "@babel/preset-react": "^7.12.10",
     "@material-ui/icons": "^4.11.2",
@@ -37,6 +38,7 @@
   "dependencies": {
     "@material-ui/core": "^4.11.2",
     "js-cookie": "^2.2.1",
-    "react-router-dom": "^5.2.0"
+    "react-router-dom": "^5.2.0",
+    "react-use": "^17.2.4"
   }
 }
